requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
greenlet==3.4.0
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pillow==12.2.0
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
six==1.17.0
SQLAlchemy==2.0.49
typing_extensions==4.15.0
typst==0.14.8
